+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fix parsing of the `quota` command output when there are devices with long names
+
 
 ## 2.3.0 - 2022-11-05